+++ src/rendering.h
@@ -11,7 +11,7 @@
 void fillRect(SDL_Renderer* renderer, const Rect& r, const Color& c);
 void drawRect(SDL_Renderer* renderer, const Rect& r, const Color& c);
 
-void drawImage(SDL_Renderer* renderer, const std::string& textureName, const vec::vec2f& position, int scale = 1);
+void drawImage(SDL_Renderer* renderer, const std::string& textureName, const vec::vec2f& position, float scale = 1);
 
 
 #endif
